adds cudnn option, runtime fallback
